Moved some serialization things to backend.
